requirements.txt
-i https://pypi.org/simple
aiobotocore==2.15.0
aiohappyeyeballs==2.4.0; python_version >= '3.8'
aiohttp==3.10.5
aioitertools==0.12.0; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
async-timeout==4.0.3; python_version < '3.11'
attrs==24.2.0; python_version >= '3.7'
blinker==1.8.2; python_version >= '3.8'
boto3==1.35.16
botocore==1.35.16; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.2.2; python_version < '3.11'
fastapi==0.114.1
flask==3.0.3
frozenlist==1.4.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
idna==3.8; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
pydantic==2.9.1; python_version >= '3.8'
pydantic-core==2.23.3; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1
s3transfer==0.10.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.38.5; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.2; python_version >= '3.10'
uvicorn==0.30.6
werkzeug==3.0.4; python_version >= '3.8'
wrapt==1.16.0; python_version >= '3.6'
yarl==1.11.1; python_version >= '3.8'
